meetme-backend/backend/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.2.0
django-oauth-toolkit==1.2.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
pytz==2018.4
requests==2.19.1
urllib3==1.23
